Flow Refactor
Move hardcoded Flow cmds to their own folder. This is a step towards refactoring them out completely.
