+++ app/commands/config.go
@@ -13,10 +13,12 @@ import (
 	"syscall"
 
 	"github.com/codegangsta/cli"
+	"github.com/codelingo/lingo/app/commands/verify"
 	"github.com/codelingo/lingo/app/util"
 	commonConfig "github.com/codelingo/lingo/app/util/common/config"
 	serviceConfig "github.com/codelingo/lingo/service/config"
 	"github.com/howeyc/gopass"
+
 	"github.com/juju/errors"
 	"golang.org/x/crypto/ssh/terminal"
 )
@@ -75,7 +77,7 @@ func init() {
 				Usage: "Set the IP of the OnPrem Platform.",
 			},
 		},
-	}, false, false, homeRq, configRq)
+	}, false, false, verify.HomeRq, verify.ConfigRq)
 }
 
 func configAction(ctx *cli.Context) {
