+++ app/commands/flows/codemod.go
@@ -1,13 +1,13 @@
-package commands
+package flows
 
 import (
 	"context"
 	"fmt"
 	"os"
-	"strings"
 
 	"github.com/codegangsta/cli"
-	"github.com/codelingo/lingo/app/commands/review"
+	"github.com/codelingo/lingo/app/commands/flows/codemod"
+	"github.com/codelingo/lingo/app/commands/verify"
 	"github.com/codelingo/lingo/app/util"
 	"github.com/codelingo/lingo/app/util/common/config"
 	"github.com/codelingo/lingo/vcs"
@@ -15,15 +15,10 @@ import (
 	"github.com/juju/errors"
 )
 
-const (
-	vcsGit string = "git"
-	vcsP4  string = "perforce"
-)
-
-var reviewCommand = cli.Command{
-	Name:        "review",
-	Usage:       "Review code following tenets in codelingo.yaml.",
-	Subcommands: cli.Commands{*pullRequestCmd},
+var CodemodCommand = cli.Command{
+	Name:        "codemod",
+	Usage:       "Modify code following tenets in codelingo.yaml.",
+	Subcommands: cli.Commands{PullRequestCmd},
 	Flags: []cli.Flag{
 		cli.StringFlag{
 			Name:  util.LingoFile.String(),
@@ -48,7 +43,7 @@ var reviewCommand = cli.Command{
 		},
 		cli.StringFlag{
 			Name:  util.DirectoryFlg.String(),
-			Usage: "Review a given directory.",
+			Usage: "Modify a given directory.",
 		},
 		cli.BoolFlag{
 			Name:  "debug",
@@ -66,21 +61,21 @@ var reviewCommand = cli.Command{
 	// "$ lingo review" will review any unstaged changes from pwd down.
 	// "$ lingo review [<filename>]" will review any unstaged changes in the named files.
 	// "$ lingo review --all [<filename>]" will review all code in the named files.
-	Action: reviewAction,
+	Action: codemodAction,
 }
 
-func reviewAction(ctx *cli.Context) {
-	err := reviewRequire()
+func codemodAction(ctx *cli.Context) {
+	err := codemodRequire()
 	if err != nil {
 		util.FatalOSErr(err)
 		return
 	}
 
-	msg, err := reviewCMD(ctx)
+	msg, err := codemodCMD(ctx)
 	if err != nil {
 		if ctx.IsSet("debug") {
 			// Debugging
-			util.Logger.Debugw("reviewAction", "err_stack", errors.ErrorStack(err))
+			util.Logger.Debugw("codemodAction", "err_stack", errors.ErrorStack(err))
 		}
 		util.FatalOSErr(err)
 		return
@@ -89,8 +84,8 @@ func reviewAction(ctx *cli.Context) {
 	fmt.Println(msg)
 }
 
-func reviewRequire() error {
-	reqs := []require{vcsRq, homeRq, authRq, configRq, versionRq}
+func codemodRequire() error {
+	reqs := []verify.Require{verify.VCSRq, verify.HomeRq, verify.AuthRq, verify.ConfigRq, verify.VersionRq}
 	for _, req := range reqs {
 		err := req.Verify()
 		if err != nil {
@@ -100,10 +95,10 @@ func reviewRequire() error {
 	return nil
 }
 
-func reviewCMD(cliCtx *cli.Context) (string, error) {
+func codemodCMD(cliCtx *cli.Context) (string, error) {
 	defer util.Logger.Sync()
 	if cliCtx.IsSet("debug") {
-		util.Logger.Debugw("reviewCMD called")
+		util.Logger.Debugw("codemodCMD called")
 	}
 	dir := cliCtx.String("directory")
 	if dir != "" {
@@ -112,7 +107,7 @@ func reviewCMD(cliCtx *cli.Context) (string, error) {
 		}
 	}
 
-	dotlingo, err := review.ReadDotLingo(cliCtx)
+	dotlingo, err := codemod.ReadDotLingo(cliCtx)
 	if err != nil {
 		return "", errors.Trace(err)
 	}
@@ -144,13 +139,6 @@ func reviewCMD(cliCtx *cli.Context) (string, error) {
 	patches, err := repo.Patches()
 	if err != nil {
 		return "", errors.Trace(err)
-	}
-	var patchesSize int64
-	for _, patch := range patches {
-		patchesSize += int64(len([]byte(patch)))
-	}
-	if patchesSize >= 1024*1024 { // >= 1MB; default max GRPC msg size accepted by the servers is 4MB.
-		util.UserFacingWarning("Warning: large diffs can be error prone. You may need to commit your changes.")
 	}
 
 	workingDir, err := repo.WorkingDir()
@@ -216,13 +204,13 @@ func reviewCMD(cliCtx *cli.Context) (string, error) {
 		return "", errors.Errorf("Invalid VCS '%s'", vcsTypeStr)
 	}
 
-	fmt.Println("Running review flow...")
-	issuec, errorc, err = review.RequestReview(ctx, req)
+	fmt.Println("Running codemod flow...")
+	issuec, errorc, err = codemod.RequestReview(ctx, req)
 	if err != nil {
 		return "", errors.Trace(err)
 	}
 
-	issues, err := review.ConfirmIssues(cancel, issuec, errorc, cliCtx.Bool("keep-all"), cliCtx.String("output"))
+	issues, err := codemod.ConfirmIssues(cancel, issuec, errorc, cliCtx.Bool("keep-all"), cliCtx.String("output"))
 	if err != nil {
 		return "", errors.Trace(err)
 	}
@@ -232,22 +220,11 @@ func reviewCMD(cliCtx *cli.Context) (string, error) {
 	}
 
 	// Remove dicarded issues from report
-	keptIssues := []*flow.Issue{}
+	var keptIssues []*codemod.SRCHunk
 	for _, issue := range issues {
 		if !issue.Discard {
 			keptIssues = append(keptIssues, issue)
 		}
 	}
-
-	// TODO(waigani) send back all issues and capture false positives.
-
-	msg, err := review.MakeReport(keptIssues, cliCtx.String("format"), cliCtx.String("output"))
-	return msg, errors.Trace(err)
-}
-
-const noCommitErrMsg = "This looks like a new repository. Please make an initial commit before running `lingo review`. This is only required for the initial commit, subsequent changes to your repo will be picked up by lingo without committing."
-
-// TODO(waigani) use typed error
-func noCommitErr(err error) bool {
-	return strings.Contains(err.Error(), "ambiguous argument 'HEAD'")
+	return "", errors.Trace(codemod.Write(keptIssues))
 }
