+++ app/commands/flows/review.go
@@ -1,12 +1,14 @@
-package commands
+package flows
 
 import (
 	"context"
 	"fmt"
 	"os"
+	"strings"
 
 	"github.com/codegangsta/cli"
-	"github.com/codelingo/lingo/app/commands/codemod"
+	"github.com/codelingo/lingo/app/commands/flows/review"
+	"github.com/codelingo/lingo/app/commands/verify"
 	"github.com/codelingo/lingo/app/util"
 	"github.com/codelingo/lingo/app/util/common/config"
 	"github.com/codelingo/lingo/vcs"
@@ -14,10 +16,15 @@ import (
 	"github.com/juju/errors"
 )
 
-var codemodCommand = cli.Command{
-	Name:        "codemod",
-	Usage:       "Modify code following tenets in codelingo.yaml.",
-	Subcommands: cli.Commands{*pullRequestCmd},
+const (
+	vcsGit string = "git"
+	vcsP4  string = "perforce"
+)
+
+var ReviewCommand = cli.Command{
+	Name:        "review",
+	Usage:       "Review code following tenets in codelingo.yaml.",
+	Subcommands: cli.Commands{PullRequestCmd},
 	Flags: []cli.Flag{
 		cli.StringFlag{
 			Name:  util.LingoFile.String(),
@@ -42,7 +49,7 @@ var codemodCommand = cli.Command{
 		},
 		cli.StringFlag{
 			Name:  util.DirectoryFlg.String(),
-			Usage: "Modify a given directory.",
+			Usage: "Review a given directory.",
 		},
 		cli.BoolFlag{
 			Name:  "debug",
@@ -60,21 +67,21 @@ var codemodCommand = cli.Command{
 	// "$ lingo review" will review any unstaged changes from pwd down.
 	// "$ lingo review [<filename>]" will review any unstaged changes in the named files.
 	// "$ lingo review --all [<filename>]" will review all code in the named files.
-	Action: codemodAction,
+	Action: reviewAction,
 }
 
-func codemodAction(ctx *cli.Context) {
-	err := codemodRequire()
+func reviewAction(ctx *cli.Context) {
+	err := reviewRequire()
 	if err != nil {
 		util.FatalOSErr(err)
 		return
 	}
 
-	msg, err := codemodCMD(ctx)
+	msg, err := reviewCMD(ctx)
 	if err != nil {
 		if ctx.IsSet("debug") {
 			// Debugging
-			util.Logger.Debugw("codemodAction", "err_stack", errors.ErrorStack(err))
+			util.Logger.Debugw("reviewAction", "err_stack", errors.ErrorStack(err))
 		}
 		util.FatalOSErr(err)
 		return
@@ -83,8 +90,8 @@ func codemodAction(ctx *cli.Context) {
 	fmt.Println(msg)
 }
 
-func codemodRequire() error {
-	reqs := []require{vcsRq, homeRq, authRq, configRq, versionRq}
+func reviewRequire() error {
+	reqs := []verify.Require{verify.VCSRq, verify.HomeRq, verify.AuthRq, verify.ConfigRq, verify.VersionRq}
 	for _, req := range reqs {
 		err := req.Verify()
 		if err != nil {
@@ -94,10 +101,10 @@ func codemodRequire() error {
 	return nil
 }
 
-func codemodCMD(cliCtx *cli.Context) (string, error) {
+func reviewCMD(cliCtx *cli.Context) (string, error) {
 	defer util.Logger.Sync()
 	if cliCtx.IsSet("debug") {
-		util.Logger.Debugw("codemodCMD called")
+		util.Logger.Debugw("reviewCMD called")
 	}
 	dir := cliCtx.String("directory")
 	if dir != "" {
@@ -106,7 +113,7 @@ func codemodCMD(cliCtx *cli.Context) (string, error) {
 		}
 	}
 
-	dotlingo, err := codemod.ReadDotLingo(cliCtx)
+	dotlingo, err := review.ReadDotLingo(cliCtx)
 	if err != nil {
 		return "", errors.Trace(err)
 	}
@@ -138,6 +145,13 @@ func codemodCMD(cliCtx *cli.Context) (string, error) {
 	patches, err := repo.Patches()
 	if err != nil {
 		return "", errors.Trace(err)
+	}
+	var patchesSize int64
+	for _, patch := range patches {
+		patchesSize += int64(len([]byte(patch)))
+	}
+	if patchesSize >= 1024*1024 { // >= 1MB; default max GRPC msg size accepted by the servers is 4MB.
+		util.UserFacingWarning("Warning: large diffs can be error prone. You may need to commit your changes.")
 	}
 
 	workingDir, err := repo.WorkingDir()
@@ -203,13 +217,13 @@ func codemodCMD(cliCtx *cli.Context) (string, error) {
 		return "", errors.Errorf("Invalid VCS '%s'", vcsTypeStr)
 	}
 
-	fmt.Println("Running codemod flow...")
-	issuec, errorc, err = codemod.RequestReview(ctx, req)
+	fmt.Println("Running review flow...")
+	issuec, errorc, err = review.RequestReview(ctx, req)
 	if err != nil {
 		return "", errors.Trace(err)
 	}
 
-	issues, err := codemod.ConfirmIssues(cancel, issuec, errorc, cliCtx.Bool("keep-all"), cliCtx.String("output"))
+	issues, err := review.ConfirmIssues(cancel, issuec, errorc, cliCtx.Bool("keep-all"), cliCtx.String("output"))
 	if err != nil {
 		return "", errors.Trace(err)
 	}
@@ -219,11 +233,22 @@ func codemodCMD(cliCtx *cli.Context) (string, error) {
 	}
 
 	// Remove dicarded issues from report
-	var keptIssues []*codemod.SRCHunk
+	keptIssues := []*flow.Issue{}
 	for _, issue := range issues {
 		if !issue.Discard {
 			keptIssues = append(keptIssues, issue)
 		}
 	}
-	return "", errors.Trace(codemod.Write(keptIssues))
+
+	// TODO(waigani) send back all issues and capture false positives.
+
+	msg, err := review.MakeReport(keptIssues, cliCtx.String("format"), cliCtx.String("output"))
+	return msg, errors.Trace(err)
+}
+
+const noCommitErrMsg = "This looks like a new repository. Please make an initial commit before running `lingo review`. This is only Required for the initial commit, subsequent changes to your repo will be picked up by lingo without committing."
+
+// TODO(waigani) use typed error
+func noCommitErr(err error) bool {
+	return strings.Contains(err.Error(), "ambiguous argument 'HEAD'")
 }
