+++ app/commands/flows/review_test.go
@@ -1,4 +1,4 @@
-package commands_test
+package flows_test
 
 import (
 	"flag"
