--- new/lingo/cmd/tenets.go
@@ -1,115 +1,0 @@
-// Copyright © 2018 CODELINGO LTD hello@codelingo.io
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package cmd
-
-import (
-	"fmt"
-	"github.com/juju/errors"
-	// "github.com/waigani/xxx"
-	// "gopkg.in/yaml.v2"
-	"io/ioutil"
-	"net/http"
-	"os"
-
-	"github.com/spf13/cobra"
-)
-
-// tenetsCmd represents the tenets command
-var tenetsCmd = &cobra.Command{
-	Use:   "tenets",
-	Short: "List Tenets",
-	Long:  ``,
-	Run: func(cmd *cobra.Command, args []string) {
-		if err := listTenets(cmd, args); err != nil {
-			fmt.Fprint(os.Stderr, err.Error())
-		}
-	},
-}
-
-func init() {
-	listCmd.AddCommand(tenetsCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// tenetsCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// tenetsCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-
-	tenetsCmd.Flags().StringP("owner", "o", "", "List all tenet bundles of the given owner")
-	tenetsCmd.Flags().StringP("bundle", "b", "", "List all tenets of the given bundle")
-	tenetsCmd.Flags().StringP("name", "n", "", "Describe the named tenet")
-	tenetsCmd.Flags().BoolP("intalled", "i", false, "List Tenets installed in current project")
-}
-
-func listTenets(cmd *cobra.Command, args []string) error {
-
-	owner := cmd.Flag("owner").Value.String()
-	bundle := cmd.Flag("bundle").Value.String()
-	name := cmd.Flag("name").Value.String()
-
-	baseTenetURL := baseDiscoveryURL + "tenets"
-	url := baseTenetURL + "/lingo_tenets.yaml"
-	switch {
-	case name != "":
-
-		if owner == "" {
-			return errors.New("owner flag must be set")
-		}
-
-		if bundle == "" {
-			return errors.New("bundle flag must be set")
-		}
-		url = fmt.Sprintf("%s/%s/%s/%s/.lingo",
-			baseTenetURL, owner, bundle, name)
-	case bundle != "":
-		if owner == "" {
-			return errors.New("owner flag must be set")
-		}
-		url = fmt.Sprintf("%s/%s/%s/lingo_bundle.yaml",
-			baseTenetURL, owner, bundle)
-
-	case owner != "":
-		url = fmt.Sprintf("%s/%s/lingo_owner.yaml",
-			baseTenetURL, owner)
-	}
-	resp, err := http.Get(url)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	defer resp.Body.Close()
-	data, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	fmt.Println(string(data))
-
-	return nil
-
-	// TODO(waigani) read into struct, sort out display
-	// var out struct{}
-
-	// err = yaml.Unmarshal(data, &out)
-	// if err != nil {
-	// 	return errors.Trace(err)
-	// }
-
-	// xxx.Dump(out)
-}
